Surveyresult

+++ pages/survey/SurveyScreen.js
@@ -1,4 +1,4 @@
-import { Text, TouchableOpacity, View } from "react-native";
+import { Text, TouchableOpacity, View, Alert } from "react-native";
 import React, { useEffect, useState } from "react";
 import { styles } from "./SurveyScreenStyle";
 import { surveyQuestions } from "./SurveyQuestions";
@@ -10,34 +10,62 @@ import {
   updateDoc,
   deleteDoc,
   setDoc,
+  getDoc,
 } from "firebase/firestore";
 import Checkbox from "expo-checkbox";
 import SurveyScreenResult from "./SurveyScreenResult";
 
 const SurveyScreen = ({ route }) => {
   const [questions, setQuestions] = useState(surveyQuestions);
-  const [ques, setQues] = useState(0);
+  let [ques, setQues] = useState(0);
   const [checkedLastQuestion, setCheckedLastQuestion] = useState([]);
   const [checkedFirstQuestion, setCheckedFirstQuestion] = useState([]);
   const [showSurveyResult, setShowSurveyResult] = useState(false);
   const [loggedInUser, setLoggedInUser] = useState(route.params.user);
   const [surveyCollection, setSurveyCollection] = useState();
+  const [retakeSurvey, setRetakeSurvey] = useState(true);
+  const [dayLeft, setDayLeft] = useState(0);
+
+  const answerString = "Answer" + [ques + 1];
+  const scoreString = "Score";
 
   useEffect(() => {
     const firestore = getFirestore();
     const surveyCollection = doc(firestore, "Survey DB", loggedInUser?.uid);
-    setSurveyCollection(surveyCollection);
-    setDoc(doc(firestore, "Survey DB", loggedInUser?.uid), {});
-  }, [loggedInUser]);
 
-  const answerString = "Answer" + [ques + 1];
-  const scoreString = "Score";
+    getDoc(surveyCollection).then((snapshot) => {
+      if (snapshot?._document !== undefined && snapshot?._document !== null) {
+        let endSurveyDate = new Date(
+          snapshot?._document?.data?.value?.mapValue.fields?.retakeDate?.timestampValue
+        );
+        const loggedInDate = new Date(loggedInUser?.metadata.lastSignInTime);
+        if (dateDiff(endSurveyDate, loggedInDate) < 14) {
+          setRetakeSurvey(false);
+        }
+      } else {
+        setRetakeSurvey(true);
+        const surveyCollection = doc(firestore, "Survey DB", loggedInUser?.uid);
+        setSurveyCollection(surveyCollection);
+        setDoc(doc(firestore, "Survey DB", loggedInUser?.uid), {});
+      }
+    });
+  }, [retakeSurvey, loggedInUser]);
+
+  useEffect(() => {
+    if (ques === 10) {
+      const retakeDate = new Date();
+
+      updateDoc(surveyCollection, {
+        retakeDate: retakeDate,
+      });
+    }
+  }, [ques]);
 
   const handleNextQuestion = () => {
     setQues(ques + 1);
   };
 
-  const handleFirestoreAnswerYes = (score) => {
+  const handleFirestoreAnswerYes = () => {
     updateDoc(surveyCollection, {
       [answerString]: {
         [answerString]: true,
@@ -47,13 +75,13 @@ const SurveyScreen = ({ route }) => {
     setQues(ques + 1);
   };
 
-  const handleFirestoreAnswerNo = (score) => {
-    /*  updateDoc(surveyCollection, {
+  const handleFirestoreAnswerNo = () => {
+    updateDoc(surveyCollection, {
       [answerString]: {
         [answerString]: false,
-        [scoreString]: questions[ques].score,
+        [scoreString]: 0,
       },
-    }); */
+    });
     setQues(ques + 1);
   };
 
@@ -72,18 +100,54 @@ const SurveyScreen = ({ route }) => {
   };
 
   const handleRetakeSurvey = () => {
-    setQues(0);
-    setCheckedLastQuestion([false, false, false, false, false]);
-    setCheckedFirstQuestion([false, false, false]);
-    setShowSurveyResult(false);
-
     const firestore = getFirestore();
+    const surveyCollection = doc(firestore, "Survey DB", loggedInUser?.uid);
 
-    deleteDoc(surveyCollection)
-      .then(() => console.log("Document deleted"))
-      .catch((error) => console.error("Error deleting document", error));
+    getDoc(surveyCollection).then((snapshot) => {
+      let endSurveyDate = new Date(
+        snapshot._document.data.value.mapValue.fields.retakeDate?.timestampValue
+      );
+      const loggedInDate = new Date(loggedInUser?.metadata.lastSignInTime);
 
-    setDoc(doc(firestore, "Survey DB", loggedInUser?.uid), {});
+      if (dateDiff(endSurveyDate, loggedInDate) > 14) {
+        setRetakeSurvey(true);
+        setQues(0);
+        setCheckedLastQuestion([false, false, false, false, false]);
+        setCheckedFirstQuestion([false, false, false]);
+        setShowSurveyResult(false);
+
+        deleteDoc(surveyCollection)
+          .then(() => console.log("Document deleted"))
+          .catch((error) => console.error("Error deleting document", error));
+
+        setDoc(doc(firestore, "Survey DB", loggedInUser?.uid), {});
+      } else {
+        setRetakeSurvey(false);
+        Alert.alert(
+          "Please Wait!",
+          `Please wait ${dayLeft} days to retake the survey`,
+          [
+            {
+              text: "Cancel",
+              style: "Cancel",
+            },
+            { text: "Okay" },
+          ]
+        );
+      }
+    });
+  };
+
+  const dateDiff = (date1, date2) => {
+    let diffTime = 0;
+    if (date1 > date2) {
+      diffTime = Math.abs(date1 - date2);
+    } else {
+      diffTime = Math.abs(date2 - date1);
+    }
+    const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
+    setDayLeft(14 - diffDays);
+    return diffDays - 1;
   };
 
   const handleSeeResults = () => {
@@ -93,7 +157,7 @@ const SurveyScreen = ({ route }) => {
   return (
     <>
       <View style={styles.container}>
-        {questions.length !== ques ? (
+        {questions.length !== ques && retakeSurvey ? (
           <View style={styles.parent}>
             <View style={styles.top}>
               <Text style={styles.questionTitle}>Question {ques + 1}</Text>
